--- HeightmapVisualizer/Program.cs
@@ -1,18 +1,0 @@
-namespace HeightmapVisualizer
-{
-    internal static class Program
-    {
-        /// <summary>
-        ///  The main entry point for the application.
-        /// </summary>
-        [STAThread]
-        static void Main()
-        {
-            // To customize application configuration such as set high DPI settings or default font,
-            // see https://aka.ms/applicationconfiguration.
-
-            ApplicationConfiguration.Initialize();
-            Application.Run(new Window());
-        }
-    }
-}
