+++ HeightmapVisualizer/src/Components/IComponent.cs
