+++ HeightmapVisualizer/src/Components/ScriptableComponent.cs
