+++ HeightmapVisualizer/src/Controls/KeyboardHandler.cs
