+++ HeightmapVisualizer/src/Primitives/DrawingMode.cs
