+++ HeightmapVisualizer/src/Primitives/Mesh.cs
