+++ HeightmapVisualizer/src/Primitives/Primitive.cs
