+++ HeightmapVisualizer/src/Window.resx
